pastas atualizadas com a versão da realse

+++ src/main/java/school/sptech/backend/service/metrica/dto/MetricaAtualizacaoDto.java
@@ -21,5 +21,5 @@ public class MetricaAtualizacaoDto {
 
     @NotNull
     @Positive
-    private int fk_usuario;
+    private long fk_usuario;
 }
